Move routing into its own module
