+++ src/article.rs
@@ -1,4 +1,4 @@
 pub mod storage;
 pub mod view;
 pub mod builder;
-
+pub use storage::gather_fs_articles;
